abbreviated months

+++ app/blog/utils.ts
@@ -76,8 +76,9 @@ export function formatDate(date: string, includeRelative = false) {
     formattedDate = 'Today'
   }
 
+  // Use 'short' for abbreviated month names
   let fullDate = targetDate.toLocaleString('en-us', {
-    month: 'long',
+    month: 'short', // Abbreviated month names (e.g., Dec)
     day: 'numeric',
     year: 'numeric',
   })
